Remove extensions submodule, move aux functions elsewhere
