--- cozy/extensions/is_same_book.py
@@ -1,2 +1,0 @@
-def is_same_book(book_a: str, book_b: str) -> bool:
-    return book_a.casefold() == book_b.casefold()
