+++ cozy/view_model/library_view_model.py
@@ -9,11 +9,10 @@ from cozy.application_settings import ApplicationSettings
 from cozy.architecture.event_sender import EventSender
 from cozy.architecture.observable import Observable
 from cozy.control.filesystem_monitor import FilesystemMonitor
-from cozy.extensions.set import split_strings_to_set
 from cozy.media.importer import Importer, ScanStatus
 from cozy.media.player import Player
 from cozy.model.book import Book
-from cozy.model.library import Library
+from cozy.model.library import Library, split_strings_to_set
 from cozy.open_view import OpenView
 from cozy.report import reporter
 from cozy.ui.import_failed_dialog import ImportFailedDialog
