+++ cozy/view_model/search_view_model.py
@@ -7,9 +7,8 @@ from cozy.application_settings import ApplicationSettings
 from cozy.architecture.event_sender import EventSender
 from cozy.architecture.observable import Observable
 from cozy.control.filesystem_monitor import FilesystemMonitor
-from cozy.extensions.set import split_strings_to_set
 from cozy.model.book import Book
-from cozy.model.library import Library
+from cozy.model.library import Library, split_strings_to_set
 from cozy.open_view import OpenView
