+++ test/cozy/model/test_library.py
@@ -28,7 +28,7 @@ def test_library_contains_books():
 
 
 def test_authors_contains_every_author_from_db():
-    from cozy.model.library import Library
+    from cozy.model.library import Library, split_strings_to_set
     from cozy.db.book import Book
 
     library = Library()
@@ -42,7 +42,7 @@ def test_authors_contains_every_author_from_db():
 
 
 def test_readers_contains_every_reader_from_db():
-    from cozy.model.library import Library
+    from cozy.model.library import Library, split_strings_to_set
     from cozy.db.book import Book
 
     library = Library()
